+++ class-table-inheritance/pom.xml
@@ -13,11 +13,7 @@
 
 
     <dependencies>
-        <dependency>
-            <groupId>checkstyle</groupId>
-            <artifactId>checkstyle</artifactId>
-            <version>3.1</version>
-        </dependency>
+
 
         <dependency>
             <groupId>org.projectlombok</groupId>
