+++ class-table-inheritance/src/main/java/com/iluwatar/classtableinheritance/Cricketer.java
@@ -1,0 +1,17 @@
+package com.iluwatar.classtableinheritance;
+
+import javax.persistence.Entity;
+import javax.persistence.Table;
+import lombok.Data;
+
+
+/**
+ * The class Cricketer is one member of class Player.
+ */
+@Data
+@Entity
+@Table(name = "cricketer")
+public class Cricketer extends Player {
+  double battingAvarage;
+
+}
