+++ class-table-inheritance/src/main/java/com/iluwatar/classtableinheritance/test.png
