--- class-table-inheritance/src/main/java/entity/Bowler.java
@@ -1,15 +1,0 @@
-package entity;
-
-import javax.persistence.Entity;
-import javax.persistence.Table;
-import lombok.Data;
-
-/**
- * Bowler belongs to one member of Cricketer.
- */
-@Entity
-@Data
-@Table(name = "bowler")
-public class Bowler extends Cricketer {
-  double bowlingAvarage;
-}
