--- class-table-inheritance/src/main/java/entity/Cricketer.java
@@ -1,17 +1,0 @@
-package entity;
-
-import javax.persistence.Entity;
-import javax.persistence.Table;
-import lombok.Data;
-
-
-/**
- * The class Cricketer is one member of class Player.
- */
-@Data
-@Entity
-@Table(name = "cricketer")
-public class Cricketer extends Player {
-  double battingAvarage;
-
-}
