+++ class-table-inheritance/src/main/resources/mapper/Player.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="mapper.MapperPlayer">
+<mapper namespace="com.iluwatar.classtableinheritance.MapperPlayer">
     <insert id="insert_player">insert into player(name) values (#{name})</insert>
     <insert id="insert_footballer">insert into footballer(name,club) values (#{name},#{club})</insert>
     <insert id="insert_Cricketer">insert into cricketer(battingAvarage,name) values (#{battingAvarage},#{name})</insert>
@@ -19,12 +19,12 @@
     <delete id="delete_Cricketer">DELETE FROM cricketer WHERE battingAvarage = #{id}</delete>
     <delete id="delete_Bowler">DELETE FROM bowler WHERE bowlingAvarage = #{id}</delete>
 
-    <select id="list_player" resultType="entity.Player">select * from player</select>
-    <select id="list_football_player" resultType="entity.Footballer">select * from footballer</select>
-    <select id="list_cricketer" resultType="entity.Cricketer">select * from cricketer</select>
-    <select id="list_bowler" resultType="entity.Bowler">select * from bowler</select>
-    <select id="get_player" resultType="entity.Player">select * from player where name=#{name}</select>
-    <select id="get_footballer" resultType="entity.Footballer">select * from footballer where name=#{name}</select>
-    <select id="get_Cricketer" resultType="entity.Cricketer">select * from cricketer where battingAvarage=#{id}</select>
-    <select id="get_Bowler" resultType="entity.Bowler">select * from bowler where bowlingAvarage=#{id}</select>
+    <select id="list_player" resultType="com.iluwatar.classtableinheritance.Player">select * from player</select>
+    <select id="list_football_player" resultType="com.iluwatar.classtableinheritance.Footballer">select * from footballer</select>
+    <select id="list_cricketer" resultType="com.iluwatar.classtableinheritance.Cricketer">select * from cricketer</select>
+    <select id="list_bowler" resultType="com.iluwatar.classtableinheritance.Bowler">select * from bowler</select>
+    <select id="get_player" resultType="com.iluwatar.classtableinheritance.Player">select * from player where name=#{name}</select>
+    <select id="get_footballer" resultType="com.iluwatar.classtableinheritance.Footballer">select * from footballer where name=#{name}</select>
+    <select id="get_Cricketer" resultType="com.iluwatar.classtableinheritance.Cricketer">select * from cricketer where battingAvarage=#{id}</select>
+    <select id="get_Bowler" resultType="com.iluwatar.classtableinheritance.Bowler">select * from bowler where bowlingAvarage=#{id}</select>
 </mapper>
